Ported to 2023 WPILib version
The 2023 NetworkTables was completely redone and API changed.
The cougar NetworkTablesTestUtil no longer works and am having trouble
figuring out how to create a pair where the client syncs with the server
so am just taking it out for the time being. That means that it isnt possible
to write a unit test for the network table control of the logger, so the logger
feature where it can be controlled by NetworkTables may not work. However the
other logger features are fine.